feat: upgrade iced dependencies from 0.13 to 0.14 (crates.io)

+++ src/application/state.rs
@@ -4,7 +4,7 @@ use crate::application::{Appearance, Application, DefaultStyle};
 use crate::core::mouse;
 use crate::core::{Color, Size};
 use crate::graphics::Viewport;
-use crate::runtime::Debug;
+use crate::debug::Debug;
 
 use std::marker::PhantomData;
 
@@ -115,7 +115,7 @@ where
                         window_info.physical_size().width,
                         window_info.physical_size().height,
                     ),
-                    scale,
+                    scale as f32,
                 );
 
                 self.viewport_version = self.viewport_version.wrapping_add(1);
@@ -173,7 +173,7 @@ where
 
                     self.viewport = Viewport::with_physical_size(
                         self.viewport.physical_size(),
-                        self.system_scale_factor,
+                        self.system_scale_factor as f32,
                     );
 
                     self.viewport_version = self.viewport_version.wrapping_add(1);
@@ -191,7 +191,7 @@ where
                     self.scale_policy = WindowScalePolicy::ScaleFactor(*new_scale);
 
                     self.viewport =
-                        Viewport::with_physical_size(self.viewport.physical_size(), *new_scale);
+                        Viewport::with_physical_size(self.viewport.physical_size(), *new_scale as f32);
 
                     self.viewport_version = self.viewport_version.wrapping_add(1);
                 }
